src/jwt/jwt.service.ts
// src/jwt/jwt.service.ts
import { Injectable } from '@nestjs/common';
import { JwtService as NestJwtService } from '@nestjs/jwt';
import { JwtPayload } from './interfaces/jwt-payload.interface';

@Injectable()
export class JwtService {
  constructor(private readonly jwtService: NestJwtService) {}

  async createToken(payload: JwtPayload): Promise<string> {
    return this.jwtService.sign(payload);
  }

  async verifyToken(token: string): Promise<any> {
    return this.jwtService.verify(token);
  }

  async refreshToken(token: string): Promise<string> {
    const payload = await this.verifyToken(token);
    delete payload.iat;
    delete payload.exp;
    return this.createToken(payload);
  }
}